encode_int to accept any positive int64

+++ include/h2o/hpack.h
@@ -34,7 +34,7 @@ extern const char *h2o_hpack_soft_err_found_invalid_char_in_header_value;
 /**
  * encodes an integer (maximum size of the output excluding the first octet is H2O_HTTP2_ENCODE_INT_MAX_LENGTH bytes)
  */
-uint8_t *h2o_hpack_encode_int(uint8_t *dst, uint32_t value, size_t prefix_bits);
+uint8_t *h2o_hpack_encode_int(uint8_t *dst, int64_t value, size_t prefix_bits);
 /**
  * encodes a huffman string and returns its length, or returns SIZE_MAX if the resulting string would be longer than the input
  */
